+++ src/library/DIPS.Mobile.UI/Components/Lists/CollectionView/CollectionView2.cs
@@ -3,4 +3,7 @@ namespace DIPS.Mobile.UI.Components.Lists;
 /// <summary>
 /// Uses the new handler on iOS
 /// </summary>
-public class CollectionView2 : CollectionView;
+public class CollectionView2 : CollectionView
+{
+    
+}